Providing C types breaks the compilation when the code is compiled with single preciseion (used by READAMR). One could potentially use a C matvec with single precision reals, so this change does not really break anything. Copying from Fortran to C arrays and back would not be very efficient. The assumption that Fortan and C integers are compatible could be relaxed by copying n, but we yet to find a case when that is necessary.
Former-commit-id: c3bf46934624c41340187696fd3b8b1caa5f7cfa

+++ share/Library/src/linear_solver_wrapper.f90
@@ -170,16 +170,16 @@ contains
 
     interface
        subroutine linear_wrapper_matvec_c(x_I, y_I, n)  bind(C) 
-         use iso_c_binding
-         integer(c_int ), VALUE:: n
-         real(c_double)        :: x_I(n)
-         real(c_double)        :: y_I(n)
+         integer, VALUE:: n
+         real          :: x_I(n)
+         real          :: y_I(n)
        end subroutine linear_wrapper_matvec_c
     end interface
 
-    integer(c_int), intent(in) :: n
-    real(c_double), intent(in) :: x_I(n) 
-    real(c_double), intent(out):: y_I(n)
+    integer, intent(in) :: n
+    real,    intent(in) :: x_I(n) 
+    real,    intent(out):: y_I(n)
+
     integer :: iStart
     !--------------------------------------------------------------------------
